requirements.txt
-i https://pypi.org/simple
alembic==1.12.1; python_version >= '3.7'
annotated-types==0.6.0; python_version >= '3.8'
asgiref==3.7.2
blinker==1.7.0; python_version >= '3.8'
cffi==1.16.0; python_version >= '3.8'
click==8.1.7; python_version >= '3.7'
colorama==0.4.6; sys_platform == 'win32'
cryptography==41.0.5; python_version >= '3.7'
exceptiongroup==1.1.3; python_version < '3.11'
flask[async]==2.3.1; python_version >= '3.8'
flask-login==0.6.3; python_version >= '3.7'
flask-migrate==4.0.5; python_version >= '3.6'
flask-sqlalchemy==3.1.1; python_version >= '3.8'
flask-wtf==1.2.1; python_version >= '3.8'
greenlet==3.0.1; platform_machine == 'aarch64' or (platform_machine == 'ppc64le' or (platform_machine == 'x86_64' or (platform_machine == 'amd64' or (platform_machine == 'AMD64' or (platform_machine == 'win32' or platform_machine == 'WIN32')))))
importlib-metadata==6.8.0; python_version < '3.10'
iniconfig==2.0.0; python_version >= '3.7'
itsdangerous==2.1.2; python_version >= '3.7'
jinja2==3.1.2; python_version >= '3.7'
mako==1.3.0; python_version >= '3.8'
markupsafe==2.1.3; python_version >= '3.7'
marshmallow==3.20.1; python_version >= '3.8'
packaging==23.2; python_version >= '3.7'
pluggy==1.3.0; python_version >= '3.8'
psycopg2-binary==2.9.9; python_version >= '3.7'
pycparser==2.21
pydantic==2.5.1; python_version >= '3.7'
pydantic-core==2.14.3; python_version >= '3.7'
pydantic-settings==2.1.0; python_version >= '3.8'
pyjwt==2.8.0; python_version >= '3.7'
pyopenssl==23.3.0; python_version >= '3.7'
pytest==7.4.3; python_version >= '3.7'
pytest-flask==1.3.0; python_version >= '3.7'
python-dotenv==1.0.0; python_version >= '3.8'
sqlalchemy==2.0.23; python_version >= '3.7'
tomli==2.0.1; python_version < '3.11'
typing-extensions==4.8.0; python_version >= '3.8'
werkzeug==2.3.0; python_version >= '3.8'
wtforms==3.1.1; python_version >= '3.8'
zipp==3.17.0; python_version >= '3.8'
